+++ sdk/azidentity/CHANGELOG.md
@@ -9,6 +9,9 @@
 ### Bugs Fixed
 * User credential types inconsistently log access token scopes
 * `DefaultAzureCredential` skips managed identity in Azure Container Instances
+* Credentials having optional tenant IDs such as `AzureCLICredential` and
+  `InteractiveBrowserCredential` require setting `AdditionallyAllowedTenants`
+  when used with some clients
 
 ### Other Changes
 * `ChainedTokenCredential` and `DefaultAzureCredential` continue to their next